requirements.txt
fastapi
uvicorn
sqlalchemy
web3
py-solc-x
setuptools